#205: Extracted common methods and fields from *CreateFragments
